Breaks up toolbox UI into individual widgets
